rethinking solution
