+++ tests/conftest.py
@@ -30,7 +30,6 @@ from gprofiler.metadata.application_identifiers import (
     get_python_app_id,
     get_ruby_app_id,
 )
-from gprofiler.metadata.application_identifiers_java import ApplicationIdentifiersJava
 from gprofiler.metadata.enrichment import EnrichmentOptions
 from gprofiler.profiler_state import ProfilerState
 from gprofiler.profilers.java import AsyncProfiledProcess, JattachJcmdRunner
@@ -563,7 +562,7 @@ def _init_profiler() -> None:
         )
     )
 
-    ApplicationIdentifiersJava.init_java(
+    ApplicationIdentifiers.init_java(
         JattachJcmdRunner(stop_event=Event(), jattach_timeout=AsyncProfiledProcess._DEFAULT_JATTACH_TIMEOUT)
     )
     set_diagnostics(False)
